Add support for logging a unique request ID in WSGI logger
Summary:
HTTP frontends like HAProxy and Nginx often support creating a globally unique
ID per request that is forwarded on to the application via a header. If this
exists, log it on each request so that individual request logs can be matched
with their logs in the HTTP frontend for debugging.

Test Plan: Need to port one of our services to it and test on staging

Reviewers: khamidou

Reviewed By: khamidou

Differential Revision: https://phab.nylas.com/D3187